Fix types for postgres server migrations

+++ migrations/server/postgres/2023-10-13-195702_create_albums/up.sql
@@ -1,11 +1,11 @@
 CREATE TABLE IF NOT EXISTS albums (
-    id INTEGER PRIMARY KEY NOT NULL,
-    artist_id INTEGER NOT NULL,
+    id BIGSERIAL PRIMARY KEY NOT NULL,
+    artist_id BIGINT NOT NULL,
     title TEXT NOT NULL,
     date_released TEXT,
     date_added TIMESTAMP NOT NULL DEFAULT NOW(),
     artwork TEXT,
     directory TEXT,
-    blur INTEGER NOT NULL DEFAULT 0,
+    blur BIGINT NOT NULL DEFAULT 0,
     FOREIGN KEY (artist_id) REFERENCES artists(id)
 )
