+++ migrations/server/postgres/2023-10-14-031701_create_tracks/up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS tracks (
-    id INTEGER PRIMARY KEY NOT NULL,
-    album_id INTEGER NOT NULL,
-    "number" INTEGER NOT NULL,
+    id BIGSERIAL PRIMARY KEY NOT NULL,
+    album_id BIGINT NOT NULL,
+    "number" BIGINT NOT NULL,
     title TEXT NOT NULL,
     duration REAL NOT NULL,
     "file" TEXT,
