+++ migrations/server/postgres/2023-10-21-224111_create_active_players/up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS active_players (
-    id INTEGER PRIMARY KEY NOT NULL,
-    session_id INTEGER NOT NULL,
-    player_id INTEGER NOT NULL,
+    id BIGSERIAL PRIMARY KEY NOT NULL,
+    session_id BIGINT NOT NULL,
+    player_id BIGINT NOT NULL,
     created TIMESTAMP NOT NULL DEFAULT NOW(),
     updated TIMESTAMP NOT NULL DEFAULT NOW()
 );
