+++ migrations/server/postgres/2024-01-16-003003_update_qobuz_config_table/up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS qobuz_config (
-    "id" INTEGER PRIMARY KEY NOT NULL,
+    "id" BIGSERIAL PRIMARY KEY NOT NULL,
     "access_token" VARCHAR(256) NOT NULL,
-    "user_id" INTEGER NOT NULL,
+    "user_id" BIGINT NOT NULL,
     "user_email" VARCHAR(256) NOT NULL,
     "user_public_id" VARCHAR(256) NOT NULL,
     created TIMESTAMP NOT NULL DEFAULT NOW(),
