+++ migrations/server/postgres/2024-01-16-003005_add_qobuz_bundle_secrets_table/up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS qobuz_bundle_secrets (
-    "id" INTEGER PRIMARY KEY NOT NULL,
-    "qobuz_bundle_id" INTEGER NOT NULL,
+    "id" BIGSERIAL PRIMARY KEY NOT NULL,
+    "qobuz_bundle_id" BIGINT NOT NULL,
     "timezone" VARCHAR(32) NOT NULL,
     "secret" VARCHAR(32) NOT NULL,
     created TIMESTAMP NOT NULL DEFAULT NOW(),
