+++ migrations/server/postgres/2024-01-22-235809_add_differing_track_type_columns_to_session_playlist_tracks_table/up.sql
@@ -2,9 +2,9 @@ ALTER TABLE session_playlist_tracks ADD COLUMN "type" VARCHAR(64) NOT NULL DEFAU
 ALTER TABLE session_playlist_tracks ADD COLUMN "data" TEXT DEFAULT NULL;
 
 CREATE TABLE IF NOT EXISTS session_playlist_tracks_temp (
-    id INTEGER PRIMARY KEY NOT NULL,
-    session_playlist_id INTEGER NOT NULL,
-    track_id INTEGER DEFAULT NULL,
+    id BIGSERIAL PRIMARY KEY NOT NULL,
+    session_playlist_id BIGINT NOT NULL,
+    track_id BIGINT DEFAULT NULL,
     "type" VARCHAR(64) NOT NULL DEFAULT 'LIBRARY',
     "data" TEXT DEFAULT NULL
 );
@@ -14,9 +14,9 @@ INSERT INTO session_playlist_tracks_temp SELECT * FROM session_playlist_tracks;
 DROP TABLE IF EXISTS session_playlist_tracks;
 
 CREATE TABLE IF NOT EXISTS session_playlist_tracks (
-    id INTEGER PRIMARY KEY NOT NULL,
-    session_playlist_id INTEGER NOT NULL,
-    track_id INTEGER DEFAULT NULL,
+    id BIGSERIAL PRIMARY KEY NOT NULL,
+    session_playlist_id BIGINT NOT NULL,
+    track_id BIGINT DEFAULT NULL,
     "type" VARCHAR(64) NOT NULL DEFAULT 'LIBRARY',
     "data" TEXT DEFAULT NULL
 );
